monster page

+++ src/view/MaximumApocalypseHelper/Monster/Monster.tsx
@@ -1,16 +1,23 @@
 import { defineComponent } from 'vue'
-import { Header } from '../../../components/Header/Header'
+import { Header } from '@/components/Header/Header'
 import { useRoute, useRouter } from 'vue-router'
 import { useMADataStore } from '../tools/dataManager'
-import { PageWrapper } from '../../../components/PageWrapper/PageWrapper'
-import { MonsterItem } from '../MonsterList/MonsterItem/MonsterItem'
+import { PageWrapper } from '@/components/PageWrapper/PageWrapper'
 import { PlayerItem } from '../PlayerList/components/PlayerItem/PlayerItem'
+import { Card } from '@/components/Card/Card'
+import s from './Monster.module.scss'
+import { MAHelperHp } from '@/view/MaximumApocalypseHelper/components/MAHelperHp/MAHelperHp'
+import { PlayerItemTheme } from '@/view/MaximumApocalypseHelper/PlayerList/components/PlayerItem/PlayerItem.interface'
+import { NoData } from '@/components/NoData/NoData'
+import { NoDataTheme } from '@/components/NoData/NoData.interface'
+import { useModal } from '@/stores/modal'
 
 export const Monster = defineComponent({
   setup() {
     const route = useRoute()
     const router = useRouter()
     const dataStore = useMADataStore()
+    const modal = useModal()
     const maDataWrapper = dataStore.getWrapper()
     const monsterId: string = route.query.id as string
     const monster = dataStore.getOne('monsters', monsterId)
@@ -28,37 +35,61 @@ export const Monster = defineComponent({
     }
 
     const handleDestroyMonster = () => {
-      dataStore.destroyMonster(monsterId)
-      router.back()
+      modal.showModal({
+        title: 'Confirm Action?',
+        content: `${monster!.name} will be killed, and be removed from the current monster list. This action can NOT be revoked!`,
+        onConfirm: () => {
+          dataStore.destroyMonster(monsterId)
+          modal.closeModal()
+          router.back()
+        },
+        onCancel: () => {
+          modal.closeModal()
+        }
+      })
+    }
+
+    const handleClickPlayer = (playerId: string) => {
+      router.push({
+        path: '/ma_helper/player',
+        query: {
+          id: playerId,
+        },
+      })
     }
 
     return () => <PageWrapper>
       <Header title={monster?.name || ''} onClickBack={() => router.back()}/>
-      {monster
-        ? <div>
-          <div>{monster.name}</div>
-          <div>
-            <div>HP: {monster.hp}</div>
-            <div onClick={() => editMonsterHp(-1)}>-</div>
-            <div onClick={() => editMonsterHp(+1)}>+</div>
+      <Card class={s.card}>
+        {monster
+          ? <div>
+            <div class={s.name}>{monster.name}</div>
+            <MAHelperHp hp={monster.hp} maxHp={monster.maxHp} onEditHp={editMonsterHp}/>
+            <div>ATK: <span class={s.propValue}>{monster.atk}</span></div>
+            <div>Position: <span class={s.propValue}>{hostPlayerMap?.name}</span></div>
+            <div>HostPlayer:</div>
+            <div>
+              {hostPlayer
+                ? <PlayerItem
+                  theme={PlayerItemTheme.MONSTER_DETAIL}
+                  onClickPlayer={() => handleClickPlayer(hostPlayer.id)}
+                  player={hostPlayer}/>
+                : <NoData theme={NoDataTheme.DARK}/>
+              }
+            </div>
+            <div>AllPlayers:</div>
+            <div>
+              {allPlayers.map((player) => <PlayerItem
+                theme={PlayerItemTheme.MONSTER_DETAIL}
+                player={player}
+              />) || <NoData theme={NoDataTheme.DARK} />}
+            </div>
+            <div class={s.destroy} onClick={() => handleDestroyMonster()}>Destroy Monster</div>
           </div>
-          <div>MaxHp: {monster.maxHp}</div>
-          <div>Position: {hostPlayerMap?.name}</div>
-          <div>ATK: {monster.atk}</div>
-          <div>HostPlayer: </div>
-          <div>
-            <PlayerItem player={hostPlayer}/>
-          </div>
-          <div>AllPlayers:</div>
-          <div>
-            {allPlayers.map((player) => <PlayerItem
-              player={player}
-            />)}
-          </div>
-          <div onClick={() => handleDestroyMonster()}>Destroy Monster</div>
-        </div>
-        : <div>No Data</div>
-      }
+          : <NoData />
+        }
+      </Card>
+
     </PageWrapper>
   },
 })
